added artwork cache
added check in ArtworkRunner.glet to see if the Spotify trackID of current track matches trackID of ID recorded in new trackId.txt file and if so, artwork is not re-downloaded.  In the event the current trackID does not match the recorded trackID, new artwork is downloaded and trackID in trackId.txt file is updated.