fix for Spotify app close issue with caching
Fixed issue where closing Spotify and re-opening wouldn't cause artwork to come up.  Was due to new trackId.txt cache file.  The trackID stored in this file didn't clear/change on app close which, when the app re-opened and resumed playing, didn't re-download the artwork as the trackID was the same.  On app close now, the trackId.txt file is altered so that artwork download is triggered on re-open.