use md5 checksum instead of trackId comparison
Removed use of trackId.txt file to keep track of current artwork and instead use md5 to compare checksum of default artwork against current artwork.  Added the checks to both when music is playing as well as when music is played/stopped or Spotify is closed so that copying of default artwork to current artwork is only done when the default checksum does not match the current artwork checksum.